PM-485 Moving marketStatus helper methods to src/store/utils
